Highlighter组件兼容HDRP
