+++ Assets/UTools/Example/UExample.unity
@@ -1075,7 +1075,7 @@ GameObject:
   - component: {fileID: 358862032}
   - component: {fileID: 358862031}
   - component: {fileID: 358862036}
-  - component: {fileID: 358862035}
+  - component: {fileID: 358862037}
   m_Layer: 0
   m_Name: Cube_TestHighlight
   m_TagString: Untagged
@@ -1171,24 +1171,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &358862035
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 358862030}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12fdbe56dfa60bb4fab2ab3722e91233, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  highlightColor: {r: 1, g: 0.92156863, b: 0.015686275, a: 1}
-  blinkSpeed: 3.5
-  minIntensity: 0.2
-  maxIntensity: 1
-  autoEnableEmission: 1
-  updateInterval: 2
 --- !u!114 &358862036
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1201,6 +1183,25 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bfbbc13fb6606ec458c91454b10bca8e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &358862037
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 358862030}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 12fdbe56dfa60bb4fab2ab3722e91233, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  highlightColor: {r: 1, g: 0.92156863, b: 0.015686275, a: 1}
+  blinkSpeed: 2
+  minIntensity: 0.2
+  maxIntensity: 1
+  autoEnableEmission: 1
+  updateInterval: 2
+  emissiveIntensity: 200
 --- !u!1 &371842240
 GameObject:
   m_ObjectHideFlags: 0
